PP-1479: Add transactional context to user service
This is to avoid deattached entities causing optimistic lock exceptions sporadically.

Because of this we had to move the exception handling for 409 and 400 results to the resource layer given that the transaction commit happens at the end of the service method.

We introduced a unit test to the resource for this moved functionality.

with @whpearson