PP-1615 Add service otp validation
We have an otp/validate endpoint. However, this also creates a user.
When we have a service invite otp to validate, we do not want to create a user.
This commit adds a new endpoint otp/validate/service which just does the otp
validation with no side effects. Ideally we will move over to using this for user
invites as well, with a secondary request to create user.

with @mrlumbu