PP-2032: Add database metrics to pay-adminusers
Registers some metrics and then pulls in the new data for the metrics whenever
the healthcheck is called.

If it can't do the db stats query for some reason it output a 0 into the stats_healthy
metric. We may or may not want to alert on this, down the line. I figured it was good
debugging info (so you could tell somewhat where the stats were broken)

As the pay-connector work does.