PP_1615 Add complete invite resource
In order to cleanly use invites to create entities (Users, Services),
it makes sense to have an action on the Invite resource that triggers the
transformation of the Invite into the required resources.

This commit adds an endpoint of form /`/v1/api/invites/{inviteCode}/complete`

which finds the specified invite, and creates user and service (if needed) based on
data in the invite. It then disables the invite.