PP-2688 Add endpoint to update merchant details
Added a new `PUT /v1/api/services/{serviceExternalId}/merchant-details` endpoint, because the semantics of this update are that we are sending the whole merchant details entity instead of PATCHing it.

The only non-mandatory field is `address_line2`.
The endpoint will return a `400 Bad Request` if not all mandatory fields are supplied in the payload request.

with @simad