PP-3096: Add chamber to docker startup script
To get this to run in ECS we require the docker container to be able to
start up the app with chamber to get the variables from the environment.

This commit adds a new codepath to run the app with chamber if the
`CHAMBER` env var is set, and it adds a chamber binary (copied from
pay-selfservice).

solo @tlwr