PP-3576 Support intrinsically Base32-encoded 2FA secrets
Currently, when we need to generate an OTP key (i.e. what’s stored
alongside each user in the database), we use one of the methods in
RandomIdGenerator (we’re not consistent: we use different methods
for direct new user creation versus invitations) to get a string.

However, the GoogleAuth library we use requires input secrets to be
Base32-encoded. To meet this requirement, we currently
(unconditionally) convert the string by taking its UTF-8 bytes and
then Base32-encoding them. This somewhat bizarre behaviour appears
to be used in order to match the behaviour of notp, a Node.js
package we used to use when we did OTP in self-service[1].

It would be more straightforward to just generate an OTP key that’s
already Base32-encoded. This commit adds such a method (not used
anywhere yet), deferring to the GoogleAuth library, which has
built-in functionality to do this (it’s unclear why we weren’t
using it before).

In addition, this commit makes it so that OTP code generation and
authorisation skips the somewhat bizarre encoding step if it
detects the key is already Base32-encoded (this changes the codes
generated and considered valid for some keys but since we currently
generate all codes ourselves, this is safe — except for period of a
few minutes during deployment where if we’re really unlucky a
subset of users may be unable to log in).

The ready Base32-encoded strings produced by the GoogleAuth library
use 80 bits of randomness to produce a sixteen-character string.
This is shorter and with less randomness than some of the methods
in the RandomIdGenerator class but it’s still quite a lot and most
OTP secrets issued by major sites seem to be sixteen characters
long (in Base32 form), so we’ll be in good company.

The motivation for doing this now is that we intend to add support
for 2FA apps and this will involve distributing OTP secrets outside
of our systems (thus making it incredibly difficult to change them
subsequently). If we did not do this, we would have to replicate
the somewhat bizarre encoding step in self-service when we generate
QR codes, and if we present the secret for manual entry, it would be
too long (whereas sixteen characters is just about manageable).

[1] https://github.com/guyht/notp/blob/master/Readme.md#google-authenticator