PP-3576 Send 2FA SMS with security code from provisional OTP key
Make it so that a call to /v1/api/users/{externalId}/second-factor
with a body of {"provisional": true} causes the provisional OTP
key to be used when generating the security code sent via text
message (error if there isn’t a provisional OTP key).

Requests without a body continue to use the regular OTP key as
before.