PP-3616: Spike out publishing results to pact broker
Unfortunately custom code is needed to publish the verification to the pact
broker because the au.com.dius.pact.pactbroker.PactBrokerClientBase fails to do
so. It seems the httpclient used internally by the PactBrokerClientBase routes
https calls to localhost:1234 and we get the following in the logs:

[a.c.d.p.p.ProviderVerifier] - Failed to publish verification results - Connect
to localhost:1234 [localhost/127.0.0.1, localhost/0:0:0:0:0:0:0:1] failed:
Connection refused (Connection refused)

This does not happen if the pact broker is using http.

solo @oswaldquek