PP-3615: A passing contract test on the provider side
For this commit:
1. Add provisional otp key when setting up test user
2. Include id data in the @State, ie. "a user exists with the given external id
7d19aff33f8948deb97ed16b2912dcd3". Unfortunately using key-value data
associated with a provider state is from the pact V3 spec. Selfservice uses the
pact-js which only supports the pact V2 spec so we have to include the external
id in the State description :(
3. For UsersApiContractTest to run locally the env vars PACT_CONSUMER_TAG,
PACT_BROKER_USERNAME and PACT_BROKER_PASSWORD must be set. For a PR build
Jenkins will have already set these.

solo: @oswaldquek