BAU: Add services/list endpoint
This endpoint gets all the services.

IMO It should be implemented as:
```
/services
```

but instead it has to be implemented as:
```
/services/list
```
(we should never have a service id which is `list`)

This is because the current `/services` endpoint is a find endpoint that
returns a single service (when you pass in `gatewayAccountId` as a query
param)

I feel like the path I chose is the lesser of two evils.

solo @tlwr