PP-4096 Save `service_name` to db
- Amended resource when creating a service to add the `service_name` if
exists
- Amended resource when creating a service and there is no `service_name`
specified to add a record that matches language ENGLISH and name the same
as the `name` parameter
- Amended ServiceCreatorTest for the field, added checks for links
- Updated ServiceDaoTest to test that the new entity `ServiceNameEntity`
is successfully persisted to the database
- Switched the order in which a ServiceNameEntity is associated with a
ServiceEntity such that when we call `add` on the set, then it will call
`equals()` on ServiceNameEntity and both entities will have the same ServiceEntity.
- Removed `id` from `equals` method as two ServiceNameEntity should be
considered equal when they have the same language, name and belong to the
same service.
- Replaced lambda method calls with reference method calls
- Replaced tests in ServiceCreateTest with unit tests and mocks for the service
- Removed tests from ServiceCreateTest that were calling the database, as
the whole chain is verified with tests for resource, service and dao
- Extracted common methods/fields into a base ServiceResourceBaseTest class
- Added helper method on the JDBI database test helper to get a list of
ServiceNameEntity based on serviceId