PP-4096 Update `service_name`
- Created temporarily a new resource where we can PATCH with an array of
operations, as per RFC6902 https://tools.ietf.org/html/rfc6902
- Renamed `ServiceEntity#serviceName` to `serviceNames` to express a collection
of elements
- Updated `ServiceEntity#addServiceName` to act as an add or updated if
exists. This is necessary when applying the patch operations and in the end
the entity is persisted to the db.
- Updated `ServiceNameEntity` to be created from `ServiceUpdateRequest`,
such that when a PATCH operation is received for service name, then it
can be created successfully
- Added `ServiceUpdatedRequest#getUpdateRequests` to return a list of
updated requests from the payload, so the PATCH payload can be deserialised
into individual `ServiceUpdateRequests`
- Created a new `ServiceRequestValidator` that will be used in the new
resource to validate for the `service_name` object as well
- Renamed `ServiceRequestValidator#FIELD_NAME` not to clash with the
actual field `FIELD_SERVICE_NAME`
- Added new validation methods in `RequestValidations` that will return
an empty collection if no errors, instead of wrapping the collection into
an `Optional` and then testing if the optional is present. Unnecessary
wrapping into `Optional`
- Added unit tests for the new resource and the new validator
- Added JSON files where to keep the payload for testing the request or
the validator. It feels easier to see the JSON object in a separate file
than to create it by hand using lists, maps etc. and to hope that you got
it right first when you try to serialize it into JSON. This is a pattern
we might adopt in other places