PP-4096 Change the API for multilingual service name updates
For updating the multilingual service name, we used to expect
requests like this:

[
  {
    "op": "replace",
    "path": "service_name",
    "value": {
      "en": "Rectify your renal records"
    }
  }
]

Now we expect requests like this:

[
  {
    "op": "replace",
    "path": "service_name/en",
    "value": "Rectify your renal records"
  }
]

In addition to being more in line with JSON Patch (but note we are
still not strictly compliant), this also makes it easier to
change multiple language variants of the service name at once in a
natural fashion:

[
  {
    "op": "replace",
    "path": "service_name/en",
    "value": "Rectify your renal records"
  },
  {
    "op": "replace",
    "path": "service_name/cy",
    "value": "Addaswch eich cofnodion arennol"
  }
]

Furthermore, it makes it simple to reuse our existing validation
for service names (no longer than 50 characters), so we do that
now. We also now validate that the specified language is actually
one we support.

Although this change breaks compatibility, it’s safe to do because
nothing is using the other API yet (it’s an internal API, not a
public one).