PP-4096 Don’t include empty/blank multilingual service names
The Service object contains a serviceNames map, which basically
only exists to get returned as JSON like this:

"service_names": {
  "en": "English Service Name",
  "cy": "Welsh Service Name"
}

Change the way the serviceNames map is constructed so that a
multilingual name is not included if the name is either the empty
string or all whitespace, as if the name were not set for that
language.

This makes it easier for upstream consumers of the JSON to
determine that there is no meaningful service name for a particular
language and fall back to something sensible (such as using the
English name, which will almost always be set).