PP-4096 Allow blank multilingual service names when updating
When receiving a request to replace one of a service’s multilingual
service names, allow empty or all-whitespace strings because — in
combination with the feature that means we won’t ever return a
blank service name when someone retrieves a service — this is the
best way we have for users to remove a multilingual service name
right now.

However, don’t allow an English service name (JSON path
service_name/en) to be blank because that is mandatory.

with @danworth