BAU: Do not execute dependency-check as part of verify phase
By default, the dependency-check plugin is tied to the verify phase
(when used as a build plugin). We don’t want this; we just want to
run it manually when required. So we define a property called
dependency-check.skip, set it to true, and use this property’s value
to set the dependency-check plugin’s own skip property. To execute
the plugin manually, override the dependency-check.skip property:

$ mvn dependency-check:check -Ddependency-check.skip=false