BAU: Stop needlessly wrapping Optionals in Optionals
We have some code like this (simplified):

return inviteDao.findByCode(inviteCode).map(inviteEntity -> {
    // Do some stuff to create an InviteCompleteResponse
    return Optional.of(response);
}).orElseGet(Optional::empty);

This is silly. The Optional.map(…) method wraps what’s returned in
an Optional, so when Optional<InviteCompleteResponse> is returned,
it becomes Optional<Optional<InviteCompleteResponse>> . Then
.orElseGet(Optional::empty) unwraps the outer Optional so we just
have Optional<InviteCompleteResponse>, which is returned.

We can avoid this muppetry by changing the code to be like this:

return inviteDao.findByCode(inviteCode).map(inviteEntity -> {
    // Do some stuff to create an InviteCompleteResponse
    return response;
});

(If for some reason we absolutely had to return
Optional.of(response) from the lambda expression, we could have
used Optional.flatMap(…) instead of Optional.map(…) to avoid it
being wrapped in an another Optional, allowing us to remove the
.orElseGet(Optional::empty) bit.)