BAU: Log errors when bootstrapping the database schema
If setting up the initial database schema fails, the reason is not logged. This
is due to a missing formatting placeholder in the log message.

Instead of logging just the exception message, pass the whole exception through
to the logger.

Credit to lgtm.com for spotting this

Also remove a redundant 'throws' clause