PP-4447 Refactoring of stripe agreement request.
Some refactoring to adopt standards encouraged by the Dropwizard documentation.
- Change `ServiceResource.createStripeAgreement()` to accept a `@NotNull` and
`@Valid` `StripeAgreementRequest`. This allows Jersey and Hibernate to handle
the validation in a resuable and easy to read manner, and declutters the logic
within the Resource method.
- Throw appropriate `WebApplicationExceptions` rather than creating a `Response`
and setting the status code to that particular error.
- Changed the return type of `ServiceResource.getStripeAgreement` to
`StripeAgreement`. Recommended by DropWizard documentation to return domain types
and avoid using `Response` where possible.
- Create `StripeAgreementRequest` to permit validation of request
by Jersey and Hibernate rather than relying upon validating a String within
the method body. The `ipAddress` is held as a String rather than a more complex
type representing a URI/ipAddress for simplicity, since we store and retrieve it
as a String without any manipulation.
- Create custom hibernate validation annotation for Ip address. This can be
reused across the project or added to pay-commons. Considers both Ipv4 and Ipv6.
- Delete `StripeAgreementRequestValidator` since this is superceeded by the
new '@ValidIpAddress` validation. Remove reference and injection of this class
in various places.