PP-4660: Remove PayLoggerFactory
We are already using MDC to embed the request ID into the logging pipeline.
This can be extracted using the %X format placeholder in the logformat, so we
don't need to create proxies to the SLF4J LoggerFactory in order to intercept
logger calls.

Remove this extra layer of indirection.