PP-4660: Use same logformat for tests and runtime
We were excluding request IDs from logs during integration tests because the
request ID header is generally not set during tests.

However, with the fallback value we now have in the logformat we can use the
integration tests as a way of visually inspecting that the fallback works as
expected.