PP-5061 Don’t read from default service name
When converting from a ServiceEntity to a Service, ignore the name
in the ServiceEntity (which is the name column in the services
table) and use the English-language ServiceNameEntity (the name
column of the 'en' row in the service_names table) instead.

This is achieved by changes to the ServiceEntity.toService()
method.

Introduce a new ServiceName object, which encapsulates an English-
language service name (mandatory) and optionally translations of
the service name into different languages (in reality, only Welsh
right now).

The new ServiceName object requires that services must have an
English-language service name defined. All real services do but
some of the ephemeral services created by automated tests don’t, so
fix those tests by making sure a name is always provided.

Though no longer read, the name property in ServiceEntity (and
hence in the name column of the services table in the database) can
still be updated (including the behaviour where updating
service_names/en also updates the name property).

This commit introduces slight changes to what is considered to be
whitespace when ignoring blank non-English service names but this
won’t make any noticeable difference. It also changes how null
name values in ServiceNameEntitys are processed but the database
constraints already ensure we never encounter these.