PP-5061 Stop updating the legacy service name
When updating the English-language service name, stop updating the
legacy service name i.e. name on ServiceEntity, which maps to the
name column in the services table in the database.

This is safe to do because nothing ever reads from the name
property of ServiceEntity for non-test purpose any more.

This may lead to the name column in the services table in the
database becoming out of sync with the service_names table but
nothing relies on the former and we intend to remove it soon.