PP-5061 Tests for creating service with only service_name.en
Add tests to ensure that we can create a service name by only
providing "service_name": {"en": "Service Name"} without a
"name": "Service Name" property.

To be clear, this functionality already worked but was not
explicitly tested.

This commit adds a little bit of duplication in the tests but this
will be removed when we stop supporting "name": "Service Name"
in create service requests.