PP-5061 Don’t write legacy service name when creating service
Don’t write legacy service name (represented by the name property
on the ServiceEntity, which maps to the name column in the services
table in the database) when creating a new service.

Since this was the last thing that set the name property on the
ServiceEntity (and no non-test code reads it now), remove the
property and its getters and setters. Fix the tests that relied
on this property (most of these should have been updated when we
stopped writing the legacy service name when updating the service
name in commit 54fd08789f7dab7ea776727c4d96ddfe887ff244 — in
retrospect, perhaps this commit and that one should have been done
in one go).