PP-5061 Ignore "name" property of service creation requests
When processing a request to create a new service, ignore the
"name" property and only use the names from the "service_name"
property.

Only selfservice creates services and it no longer sends the
"name" property (but it does always include "service_name" with
at least an "en" property).

Since this involves changes to the ServiceResource and
ServiceCreator, make them not use an Optional<List<String>> for
the gateway account IDs because optionals of lists are cursed.