Create service: don’t hard-code recognised service name languages
When deserialising a JSON payload for creating a service, don’t
only look specifically for English and Welsh service names.
Instead, iterate over the values of the SupportedLanguage enum
and look for service names in those languages.

This further centralises the SupportedLanguage enum as the single
source of truth for what languages we support and will make adding
support for more languages in the future require fewer code
changes.