REP: Fix asynchronous calls to actually resolve
In some services where `CompletableFuture` gets used, it sets up
handlers for error cases from calls to external services but doesn't
actually wait for them to resolve before returning from the
encapsulating method. The issue with this becomes clearer in tests where
the test attempts to assert on the status of the `CompletableFuture`
before it has completed.

`.get()` should wait to complete.