REP: Stop misusing asynchronous calls to Notify
The `NotificationService` was returning things as `CompletableFuture` in
some hopes of making the calls to Notify's API asynchronous. However the
status of the calls was never checked and it looks like adminusers
doesn't reallu care whether the API call was successful or not and
carries on regardless. I've not changed this behaviour but instead
explicitly catch the `Exception`s thrown by `NotificationService` and
avoided the mess introduced by `CompletableFuture`.