Name JUnit @Before methods setUp() instead of setup()
setUp() was what JUnit required in the pre-annotation JUnit 3 days
when methods with certain names had magic behaviour. It’s nice to
stick to this convention for consistency.

Rename the setUp() method in IntegrationTest to initialise()
because other tests subclass IntegrationTest and if they both have
@Before methods with the same name, the one in the subclass
overrides the other one and causes it not to be run (this wasn’t
a problem before because setUp() and setup() are different as far
as Java is concerned).