Introduce ShellCheck
Fix or ignore errors in shell scripts found by ShellCheck.

ShellCheck reads a shellcheckrc file in the project directory,
which can contain directives like disable=SC2059 to ignore certain
errors.

Add a shellcheckrc file with no directives (so far, everything
we’ve wanted to ignore has been best dealt with on a case-by-case
basis) for the benefit of Codacy.