PP-5809: Don't throw runtime exception on failure to send notification
When there is an exception sending a notification, the
`uk.gov.service.notify.NotificationClient` will log an error. So there is no
need for us to rethrow a RuntimeException which will also be logged. This is
causing duplicate issues in Sentry. For example,

https://pay-sentry.cloudapps.digital/sentry/adminusers/issues/299/ and
https://pay-sentry.cloudapps.digital/sentry/adminusers/issues/300/ are the same
error - they have the same sentry threadname dw-25. Similarly for
https://pay-sentry.cloudapps.digital/sentry/adminusers/issues/1242/ and
https://pay-sentry.cloudapps.digital/sentry/adminusers/issues/1241/.