PP-6000 Allow choosing the Notify template for sending OTP via SMS
Change NotificationService so that when the
sendSecondFactorPasscodeSms(…) method is called, an enum constant
from the new OtpNotifySmsTemplateId enum type is passed in to
indicate which Notify template ID to use.

For now, all callers of this method pass in the LEGACY enum
constant, which maps to the long-serving Notify template ID, so
there’s no behaviour change yet.

In subsequent commits, different templates will be chosen for
different types of messages and LEGACY will be removed.

Also add unit tests for NotificationService because there weren’t
any and the logic is becoming a bit more complex now. This
required some tweaks to how NotificationService is instantiated to
make it easier to inject mock dependencies.