PP-6000 Specific templates for sign in v change sign-in method SMS
When sending an OTP code to an existing user (not a user who is in
the process of creating their account), use the new specific
Notify template IDs for the two possible activities: signing in or
changing the sign-in method to SMS.

Note that the actual contents of the messages are the same for now
but this allows them to diverge in the future (which we’ll do
when nothing is using the legacy template ID).

This change subtly changes the semantics of what it means to send
an OTP for a user. Previously, the only difference was whether the
provisional OTP key was used or not. Now this also implies whether
the user is signing in or changing their sign-in method to SMS and
affects the message content, so adjust some naming to clarify this.