PP-6000 Make InviteEntity type field an InviteType not String
Make the InviteEntity type field an InviteType and not a String.
Also make its the InviteEntity getType() method return an
InviteType not a String. Use InviteEntity’s existing isUserInvite()
and isServiceInvite() methods rather than getType() in some places.

This means that any InviteEntity that is loaded from the database
and has a type that isn’t “user” or “service” will blow up during
deserialisation but there are no such entities in the production
database and other things would probably break if there were.