PP-6000 Specific templates for different types of user creation SMS
When sending an OTP code for user who is trying to create a new
account (not a user who already has an account), use the new
specific Notify template IDs for the two possible variations:
self-initiated user and service creation or creating a user in
response to an invitation to join a service.

Note that the actual contents of the messages are the same for now
but this allows them to diverge in the future (which we’ll do
when nothing is using the legacy template ID).