PP-6085 Lookup DB creds using the service name within env-map.yml
The location of the DB credentials will change within VCAP_SERVICES if
the database is a postgres rds instance provided as a cf service
(under 'postgres'), or a container which we use in dev spaces (under
'user-provided').

'env-map.yml' should use the service name which should always be 'adminusers-db'
regardless of the environment.