PP-7458 Capture created date for services
`created_date` column already exists with a default, but as it is a
nullable column the value was set to the null value on the
ServiceEntity, as we are using JPA for entity management.

Fix this by adding a default value of `now()` on the ServiceEntity. Add
an integration test to ensure a `created_date` is set and returned.