PP-7663: Endpoint to get admin emails for gateway accounts
We've decided to return a list of admin emails per gateway account, instead of
a list User objects (which contains the email) as we'd be potentially asking for
hundreds of gateway accounts. This will mean less data transferred over the wire.

There's potential for generalising return bodies using
`@JsonView(UserEntity.Views.ToolboxView.class)` in the future if we have the
time.