PP-7964 Users: toUnmodifiableList() and toUnmodifiableMap()
Replace Collectors.toList() with Collectors.toUnmodifiableList()
in UserService and replace Collectors.toList() with
Collectors.toUnmodifiableList() and Collectors.toMap() with
Collectors.toUnmodifiableMap() in UserDao.

The returned lists are never modified and will never contain nulls.

The returned map never contains null keys or values but it was
modified in some cases, so fix that with an intermediary modifiable
map and add a unit test to make sure it works correctly.