BAU — Make integration test pass when computer time zone is BST
Make ServiceDaoIT.shouldFindByServiceExternalId() pass when the
computer time zone is British Summer Time (+0100).

The test inserts a test into the database, retrieves it using the
ServiceDao and then asserts that the created date of the service is
within 5 seconds of the current time. When the test runs on a
computer using BST as its time zone, it fails because the created
date is an hour earlier than the actual current time.

For example, if it is 15:00 BST, the test expects the created date
of the service to be 14:00 UTC (which is the same time because BST
is an hour ahead of UTC). However, the created date is 13:00 UTC.

I am not sure why this is. The service is inserted into the database
directly using JDBI. The created date is stored in a column with a
type of TIMESTAMP WITH TIME ZONE with a default value of
timezone('utc'::text, now()). The DatabaseTestHelper code that
inserts the service does not explicitly specify a created date, so
the default is used.

For some reason, when the test runs, the created date that gets
inserted into the column is an hour earlier than expected. For
example, if it’s currently 15:00 BST, 13:00 UTC gets inserted rather
than 14:00 UTC (I have peered into the database while the test is
running to verify this).

If I insert a new row into the database using the standard
PostgreSQL command line client, the behaviour is as expected: 15:00
BST becomes 14:00 UTC.

Unfortunately, I cannot figure out what’s going wrong. I presume
somewhere along the line the BST date is having an hour subtracted
to convert it to UTC but then something assumes it’s still UTC and
subtracts another hour.

Make the test pass by changing DatabaseTestHelper so that it inserts
the service into the database with the created date it already has
rather than relying on the database column default. The real code
(as opposed to the test code) always specifies the created date
itself and does not rely on the database column default so this is
reasonable. Furthermore, the production servers always run in UTC
so nothing like this is likely to happen.

To make this work, change some instances of ZoneId.of("UTC") to
ZoneOffset.UTC because they’re not actually equal to each other
(another reason we should use Instant instead of ZonedDateTime but
that’s for another time).