PP-8728 Do not require password and phone for service invite
Stop requiring a phone number and password to be specified when
we get a request from selfservice to create a service invite so
that we can eventually only set the phone number and password
on the invite after it’s created.

Currently, selfservice’s validation will stop us from ever
receiving a request to create a service invite that does not have
a phone number and a password.