PP-8728 Set phone and password on service invite when sending OTP
If a request is sent to /v1/api/invites/{code}/otp/generate with
the following body:

{
  "telephone_number": "07700900000",
  "password": "password123"
}

… and the code relates to a service invite (that is, one where the
user is signing up on their own initiative without receiving an
invitation to join a service) then update the phone number and
password on the invite with the provided values before sending the
one-time password.

This makes ServiceOtpDispatcher behave similarly to
UserOtpDispatcher (the equivalent for user invites — that is, where
a user is signing up because they have been invited to a service)
except the "telephone_number" and "password" properties are
optional for ServiceOtpDispatcher.

This prepares us for a world where service invites only have the
phone number and password set on them immediately before the one-
time password is sent rather than when they’re created.