PP-9260 Add event subscriber queue message receiver
Add a receiver Managed class which starts a ScheduledExecutorService to
poll an SQS queue for event messages and all required config for this.

The event messages expected have the same structure as event messages in
Ledger. The Event model for deserialisation only currently deserialises
the fields that we expect adminusers to consume.