PP-8827: Endpoint to remove user from service
Given that users need to be in the alphagov organisation in GitHub in order to
be authenticated with toolbox, it should be fairly safe to expose an endpoint
to remove a user from a service without requiring a "GovUkPay-User-Context"
header. Note that the original `DELETE
/v1/api/services/{serviceExternalId}/users/{userExternalId}` endpoint is still
valid and is used by selfservice. In order to make it clear `DELETE
/v1/api/toolbox/services/{serviceExternalId}/users/{userExternalId}` is for
toolbox use only we've placed the implementation in
ToolboxEndpointResource.java for clarity.