PP-9677 Remove fields from dispute lost email personalisation
Remove the disputeFee and disputeAmount fields from the personalisation
as these are not yet used by the Notify template.

The `fee` and `net_amount` fields will only be sent in the DISPUTE_LOST
event from connector when we are charging for disputes.

The story to update the template with information about the amount
charged for lost disputes can add these back in where appropriate.