PP-9292 Fix deserialisation of events
Ledger was previously re-serialising events to send to SNS with the
`event_details` value as an escaped JSON string.

This has been changed so that the event processed by ledger is directly
passed on without deserialising and then re-serialising the event
message in a way that transformed the `event_details` in this way.