PP-9292 Add contract test for DISPUTE_CREATED event
Add a contact test to verify that connector sends the DISPUTE_CREATED
event with the content expected by adminusers in order to create the
personalisation for the email to send using Notify.

This copies some of the classes for the queue message contract tests
from ledger and modifies them as appropriate.

The `queueSchedulerThreadDelayInSeconds` configuration option is changed
to `queueSchedulerThreadDelayInMilliseconds` to allow for a short delay
for tests. No infra sets the
`EVENT_SUBSCRIBER_QUEUE_SCHEDULER_THREAD_DELAY_IN_SECONDS` environment
variable, meaning this change is fine to make.

Add step to the post-merge github actions workflow to run the connector
provider tests.