PP-9814 Fix the OpenApi spec for search services
The OpenApi schema for the search services response was incorrectly
stating that the response was an array of services. However, it actually
returns an object with two fields: `name_results` and `merchant_results`

Fix this by creating a Java class to represent the response, rather than
returning a Map and use this as the schema for the OpenApi spec.

Regenerated .secrets.baseline file using version 1.2.0 of
detect-secrets, as used by other repositories.